+++ OUDS/Core/Components/Sources/Link/Internal/LinkColorIndicatorModifier.swift
@@ -12,6 +12,7 @@
 //
 
 import OUDS
+import OUDSFoundations
 import OUDSTokensComponent
 import OUDSTokensSemantic
 import SwiftUI
@@ -24,26 +25,30 @@ struct LinkColorIndicatorModifier: ViewModifier {
 
     @Environment(\.theme) private var theme
     @Environment(\.oudsUseMonochrome) private var useMonochrome
+    @Environment(\.oudsSurfaceColor) private var surfaceColor
 
     // MARK: - Body
 
     func body(content: Content) -> some View {
-        content.oudsForegroundStyle(appliedColor)
+        content.oudsForegroundStyle(appliedColor())
     }
 
     // MARK: - Helpers
 
-    private var appliedColor: MultipleColorSemanticTokens {
+    private func appliedColor() -> MultipleColorSemanticTokens {
+        let colorToApply: MultipleColorSemanticTokens
         switch interactionState {
         case .enabled:
-            enabledColor
+            colorToApply = enabledColor
         case .hover:
-            hoverColor
+            colorToApply = hoverColor
         case .pressed:
-            pressedColor
+            colorToApply = pressedColor
         case .disabled, .readOnly:
-            disabledColor
+            colorToApply = disabledColor
         }
+        OUDSWCAG21Ratios.debugContrastRatio(colorToApply, surfaceColor, .nonTextual)
+        return colorToApply
     }
 
     private var enabledColor: MultipleColorSemanticTokens {